+++ dv.jquery.js
@@ -13,7 +13,7 @@
 
   $.fn.dvHtml = function (dv) {
     var el = this;
-    dv.onchange(function (v) { el.html(v); });
+    dv.onChange(function (v) { el.html(v); });
     el.html(dv.value);
     return el;
   }
@@ -21,7 +21,7 @@
   $.fn.dvVal = function (dv) {
     var el = this;
     el.on("input", function () { dv.value = el.val(); });
-    dv.onchange(function (v) { el.val(v); });
+    dv.onChange(function (v) { el.val(v); });
     el.val(dv.value);
     return el;
   };